refactor:create component selectEvenBar
