+++ lib/components/selected_bar/events/eventSelected.dart
